requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
python-docx
python-pptx
reportlab
openai
